separate 1d and 2d indexation tests
